Add make target to build Android fonts from hinted fonts.
The change is temporary, while we confirm the quality of our new
build toolchain. The target should go away once we switch completely
to the new toolchain.

This also removes the code that removed tab and unassigned characters
from the Android target, since the tab now has the same width as
space and the unassigned characters are no longer in the fonts.