Correct family and style data in name table.
This uncomments some code which distinguishes between Regular/Bold and other
weights when setting the UFOs' styleMapFamilyName attribute, which is used by
the FDK to set the output OTFs' family names (this association is documented
here: http://unifiedfontobject.org/versions/ufo2/fontinfo.html). It's not clear
why this code was commented out in the first place.

Part of #37