Use a nicer hack for approximating cubic curves.
To convert cubic curves to quadratic splines, we split the cubic curve in half
and then approximate each half (which is itself a cubic curve) with a single
quadratic curve.

For this approximation we previously used the intersection of the cubic curve's
control vectors as the second point in its resulting quadratic curve, which
sometimes exploded into wildly inaccurate approximations. This new approximation
method is more stable.