Explicitly check anchor objects against None.
This is necessary because of some implementation details in RoboFab;
RAnchor objects always evaluate to False in a boolean context.

Fixes a bug in which "top" anchors were not being added to composite
glyphs, since they are added only when the "parenttop" anchor is
present (which was checked for with a statement like "if anchor:").