Always deal with floats when mixing font masters.
Without this change, some of the master data contains int values
which are not interpolated correctly due to numpy array behavior.

For example adding float values to a numpy array of ints will
always result in an array of ints, so a scale value which should be
a float between 0 and 1 might instead be rounded.