Provide a hotfix for webfont names.
This ensures that we'll only have four fonts with family name
"Roboto" (regular, bold, italic, and bold italic), which is necessary
for Windows compatibility. Other styles have their weight appended to
the family name.

Also ensures that the macStyle bold bit is set for weight Black, to be
consistent with older Roboto binaries.