Hotfix features to be unaffected by class sorting
This is a work-around for a bug in which class members are sorted
before compilation. Essentially you can't have something like
[w uni0100] mapping to [W uni101], because the first class will get
sorted to [uni0100 w]. We can consolidate the classes again (probably
better than before) when the bug is fixed.

This also fixes a bad smcp mapping from dotlessi to Idotaccent.