Add back chi as an alias of chi.alt
This is a partial revert of 543a1804e99e0866c9e6d89bce506330365be9e1,
and it keeps the same number of glyph as the output had before that
commit. That is, before there were glyphs "x" and "chi" which were
identical, after this there are "chi.alt" and "chi" which are
identical.

Again, this is done for consistency with the hinted binaries.