Copy features from unhinted to hinted for web
This effectively applies bf634faa082595d3c8d14ad1c02bb3b6949a1950 to
the web target, because the bug fixed by that commit is still present
in the hinted binaries.

This also removes the subset step from the web target, which we may or
may not want to keep. It can always be put back in.