Re-incorporate missing anchors into UFO sources
Before these were extracted from VFB sources separately due to a bug
in vfb2ufo. But now that we're not tracking VFB sources, they should
just be in the UFOs.

Unfortunately, the format of XML output by ufoLib is different than
the existing format. So this commit is much larger than it needs to
be. Most of the changes should just be indentation.