Remove alias and combAccentOnly flags for mark pos
The alias flags were used to apply mark positioning to a composite
glyph based on its components. But we can achieve the same effect by
simply making sure the anchors for those composite glyphs are copied
from their components.

We do this by adding the aliased glyphs' unicode values to a set which
filters the composite glyphs to which anchors are copied. I'm not sure
why this filter exists in the first place, but for now I'm keeping it
in to be safe.

The combAccentOnly flags were used to apply mark-to-base attachments
only for accent glyphs with width of zero. This was unnecessary
because no nonzero-width-glyph contains any of the mark anchors for
which the flag was applied.