Count glyphs with _bottom anchor as accents
This fixes a small regression due to a change in ufo2ft in which only
glyphs which contained one of any provided accent anchors was used as
a base in mark-to-mark positioning rules (where before any glyph could
be used as a base).

I think that change makes sense generally, but it means that we have
to include this workaround in which we provide a dummy base-accent
anchor pair simply to specify that glyphs containing the _bottom
anchor should be used as mkmk bases.