Fix composite glyph flags, #251.
The dependency ufo2ft had a bug, https://github.com/googlei18n/ufo2ft/issues/136
which set the flags of composite glyphs incorrectly.

Solving the issue is rather complex. The previous implementation attempted to
convert cubic splines to quadratic twice, once in fonts_to_quadratic and again
in compileTTF. The conversion in compileTTF also reversed the contour
direction. The composite flags were getting set incorrectly in compileTTF's
conversion. To solve it, we disable conversion in compileTTF and enable the
reverse_direction parameter in fonts_to_quadratic.

Upgrading the ufo2ft dependency was not a viable option. The interfaces for
some of its sub dependencies has changed.